requirements.txt
GitPython==3.1.43
PyGithub==2.3.0
semver==3.0.2
